refactor: make Popover more composable
